requirements.txt
asyncpg==0.29.0
grpcio==1.63.0
grpcio-tools==1.63.0
importlib_metadata==7.1.0
protobuf==5.26.1
psycopg2==2.9.9
python-dotenv==1.0.1
setuptools==69.5.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tabulate==0.9.0
typing_extensions==4.11.0
yoyo-migrations==8.2.0
zipp==3.18.1